made the methods make more sense, added ingrediants, added comments
