--- app/src/main/java/com/healthyeats/model/json/JsonWriter.java
@@ -1,110 +1,0 @@
-package com.healthyeats.model.json;
-
-
-import android.content.Context;
-import android.util.Log;
-
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-import com.healthyeats.model.recipe.Recipe;
-
-import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class JsonWriter {
-
-    public JsonWriter(Context context){
-        try {
-            OutputStreamWriter outputStreamWriter = new OutputStreamWriter(context.openFileOutput("recipes.json", Context.MODE_PRIVATE));
-            outputStreamWriter.close();
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-    public void recipeDeleter(Recipe recipe, Context context) {
-        Gson gson = new Gson();
-        String ret = "";
-        try {
-            InputStream inputStream = context.openFileInput("recipes.json");
-            if (inputStream != null) {
-                InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
-                BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
-                String receiveString = "";
-                StringBuilder stringBuilder = new StringBuilder();
-                while ((receiveString = bufferedReader.readLine()) != null) {
-                    stringBuilder.append("\n").append(receiveString);
-                }
-                inputStream.close();
-                ret = stringBuilder.toString();
-            }
-        } catch (FileNotFoundException e) {
-            Log.e("login activity", "File not found: " + e.toString());
-        } catch (IOException e) {
-            Log.e("login activity", "Can not read file: " + e.toString());
-        }
-        Type listRecipeType = new TypeToken<List<Recipe>>() {}.getType();
-        List<Recipe> recipes = gson.fromJson(ret, listRecipeType);
-        for(int i = 0; i < recipes.size(); i++){
-            if( recipes.get(i).getName().equals(recipe.getName())){
-                recipes.remove(i);
-            }
-        }
-
-
-        try {
-            OutputStreamWriter outputStreamWriter = new OutputStreamWriter(context.openFileOutput("recipes.json", Context.MODE_PRIVATE));
-            outputStreamWriter.write(gson.toJson(recipes));
-            outputStreamWriter.close();
-        } catch (IOException e) {
-            Log.e("Exception", "File write failed: " + e.toString());
-        }
-    }
-
-
-    public void writeToFile(Recipe recipe, Context context) {
-        Gson gson = new Gson();
-        String ret = "";
-        try {
-            InputStream inputStream = context.openFileInput("recipes.json");
-            if (inputStream != null) {
-                InputStreamReader inputStreamReader = new InputStreamReader(inputStream);
-                BufferedReader bufferedReader = new BufferedReader(inputStreamReader);
-                String receiveString = "";
-                StringBuilder stringBuilder = new StringBuilder();
-                while ((receiveString = bufferedReader.readLine()) != null) {
-                    stringBuilder.append("\n").append(receiveString);
-                }
-                inputStream.close();
-                ret = stringBuilder.toString();
-            }
-        } catch (FileNotFoundException e) {
-            Log.e("login activity", "File not found: " + e.toString());
-        } catch (IOException e) {
-            Log.e("login activity", "Can not read file: " + e.toString());
-        }
-        Type listRecipeType = new TypeToken<List<Recipe>>() {}.getType();
-        List<Recipe> recipes = gson.fromJson(ret, listRecipeType);
-        if(recipes == null){
-            recipes = new ArrayList<Recipe>();
-        }
-        recipes.add(recipe);
-
-        try {
-            OutputStreamWriter outputStreamWriter = new OutputStreamWriter(context.openFileOutput("recipes.json", Context.MODE_PRIVATE));
-            outputStreamWriter.write(gson.toJson(recipes));
-            outputStreamWriter.close();
-        } catch (IOException e) {
-            Log.e("Exception", "File write failed: " + e.toString());
-        }
-    }
-}
